+++ src/routes/graphql/index.ts
@@ -1,12 +1,12 @@
 import { FastifyPluginAsyncJsonSchemaToTs } from '@fastify/type-provider-json-schema-to-ts';
 import { GraphQLSchema, graphql, parse, validate } from 'graphql';
+import depthLimit = require('graphql-depth-limit');
 
 import { graphqlBodySchema } from './schema';
 import rootQuery from './query';
 import mutation from './mutations';
-import depthLimit = require('graphql-depth-limit');
 
-const QUERY_DEPTH_LIMIT = 6;
+const DEPTH_LIMIT = 6;
 
 const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
   fastify
@@ -18,7 +18,7 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
         body: graphqlBodySchema,
       },
     },
-    async function (request, reply) {
+    async function (request) {
       const source = request.body.query!;
 
       const schema = new GraphQLSchema({
@@ -27,13 +27,13 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       });
 
       const [error] = validate(schema, parse(source), [
-        depthLimit(QUERY_DEPTH_LIMIT),
+        depthLimit(DEPTH_LIMIT),
       ]);
 
       if (error) {
         return {
           errors: {
-            message: `Query exceeds maximum operation depth of ${QUERY_DEPTH_LIMIT}`,
+            message: `Query exceeds maximum operation depth of ${DEPTH_LIMIT}`,
           },
           data: null,
         };
