+++ src/routes/graphql/types/member-type.ts
@@ -1,0 +1,32 @@
+import {
+  GraphQLObjectType,
+  GraphQLString,
+  GraphQLInt,
+  GraphQLNonNull,
+  GraphQLInputObjectType,
+} from 'graphql';
+
+export const MemberType = new GraphQLObjectType({
+  name: 'MemberType',
+  fields: {
+    id: { type: new GraphQLNonNull(GraphQLString) },
+    discount: { type: new GraphQLNonNull(GraphQLInt) },
+    monthPostsLimit: { type: new GraphQLNonNull(GraphQLString) },
+  },
+});
+
+export const MemberTypeInput = new GraphQLInputObjectType({
+  name: "MemberTypeInput",
+  fields: {
+    discount: { type: new GraphQLNonNull(GraphQLInt) },
+    monthPostsLimit: { type: new GraphQLNonNull(GraphQLInt) },
+  },
+});
+
+export const MemberTypeUpdateInput = new GraphQLInputObjectType({
+  name: "MemberTypeUpdateInput",
+  fields: {
+    discount: { type: GraphQLInt },
+    monthPostsLimit: { type: GraphQLInt },
+  },
+});
